Expose worktree removal API from the library crate
Move the worktree-removal flow (fsmonitor cleanup, fast-path trash
staging, safe branch deletion) out of the binary-only `commands`
module into the library crate so external tools like `worktrunk-sync`
can call it without reimplementing with raw git commands.

The new entry point is `worktrunk::git::remove_worktree_with_cleanup`,
taking a `RemoveOptions` struct that replaces the six positional args
of the previous internal `execute_removal`. `BranchDeletionMode`,
`BranchDeletionOutcome`, `BranchDeletionResult`, and `RemovalOutput`
are re-exported from `worktrunk::git` alongside two lower-level
helpers (`stage_worktree_removal`, `delete_branch_if_safe`) for
callers that need finer control.

Internal callers (`wt remove`, `wt merge --remove`, the TUI picker)
now go through the library module — there is no duplicate definition
left in `commands/`.
