+++ src/commands/mod.rs
@@ -1,5 +1,4 @@
 mod alias;
-pub(crate) mod branch_deletion;
 pub(crate) mod command_approval;
 pub(crate) mod command_executor;
 pub(crate) mod commit;
