+++ src/git/mod.rs
@@ -8,6 +8,7 @@ mod error;
 mod parse;
 pub mod recover;
 pub mod remote_ref;
+pub mod remove;
 mod repository;
 mod url;
 
@@ -56,6 +57,10 @@ pub use error::{
 };
 pub use parse::{parse_porcelain_z, parse_untracked_files};
 pub use recover::{current_or_recover, cwd_removed_hint};
+pub use remove::{
+    BranchDeletionMode, BranchDeletionOutcome, BranchDeletionResult, RemovalOutput, RemoveOptions,
+    delete_branch_if_safe, remove_worktree_with_cleanup, stage_worktree_removal,
+};
 pub use repository::{Branch, Repository, ResolvedWorktree, WorkingTree, set_base_path};
 pub use url::GitRemoteUrl;
 pub use url::parse_owner_repo;
